fix: a bug fix
docs: change to documentation

refactor: refactoring code

The feature allowing users to move books between shelves

Change in the documentation of the Readme

Change in the constructor, method componentDidMount and updateBook of the class App. Change in the method render of the class ListShelves. Change in the method updateQuery and the search request. When the users perform a search, they can add to bookshelf without going to main page.